\name{monthlyStats}
\alias{monthlyStats}
\title{
  Monthly returns}
\description{
  compute monthly returns
}
\usage{
monthlyStats(x, t = NULL)
}
\arguments{
  \item{x}{
    a zoo series or a vector
  }
  \item{t}{
    time vector
  }

}
\details{
  computes an index and returns
}
\value{
  a list
  \item{index}{scaled performance}
  \item{returns}{monthyl returns}

}
\references{
  GMS
}
\author{
  Enrico Schumann
}
\examples{
##
}
