\name{instrument}
\alias{instrument}
\title{
  Retrieve or Change Instrument
}
\description{
  Generic function for retrieving and changing instrument information.
}
\usage{
instrument(x, \dots)
}
\arguments{
  \item{x}{
    an object
  }
  \item{\dots}{
    arguments passed to \code{\link{print}}
  }
}
\details{

  Generic function.
  
}
\value{

  A character vector.
  
}
\references{

  Schumann, E. (2016) \emph{Portfolio Management with
    R}. \url{http://enricoschumann.net/R/packages/PMwR/manual/PMwR.html}

}
\author{
  Enrico Schumann
}
\section{Warning }{
   Experimental. The interface may change.
}
\seealso{
  \code{\link{position}}
}
\examples{
## TO BE ADDED
}
