\name{PM-internal}
\alias{PM-internal}
\alias{makeHHMMSS}
\title{Internal functions in the PM package}
\description{
  Several internal functions, not exported from the \pkg{PM} name space.
}
\usage{
makeHHMMSS(x, label = "time specification (HHMMSS)")
}
\details{
  \code{makeHHMMSS} transforms the character \code{x} into a
  character vector that resembles a format \code{HHMMSS}. 
}

\author{
  Enrico Schumann
}

\keyword{internal}
