\name{PM-package}
\alias{PM-package}
\alias{PM}
\docType{package}
\title{
Functions for the management of financial portfolios
}
\description{
Functions for the management of financial portfolios.
}
\details{
There is little so say.
}
\author{
Enrico Schumann

Maintainer: Enrico Schumann <es@enricoschumann.net>
}
\references{
Gilli, M., Maringer, D. and Schumann, E. (2011) \emph{Numerical Methods and Optimization in Finance}. Elsevier.
\url{http://www.elsevierdirect.com/product.jsp?isbn=9780123756626}
}
\keyword{package}
