\name{PMwR-package}
\alias{PMwR-package}
\alias{PMwR}
\docType{package}
\title{

  Functions for the Practical Management of Financial Portfolios

}
\description{

  Functions for the practical management of financial portfolios:
computing profit-and-loss, analysing trades, reporting and more.

}
\details{

  This package is work in progress.  Functions may change or disappear
  without notice; in particular, function interfaces are not stable.
  
}
\author{
Enrico Schumann

Maintainer: Enrico Schumann <es@enricoschumann.net>
}
\references{

Gilli, M., Maringer, D. and Schumann, E. (2011) \emph{Numerical Methods
and Optimization in Finance}. Elsevier.
\url{http://www.elsevierdirect.com/product.jsp?isbn=9780123756626}

Schumann, E. (2013) \emph{Portfolio Management with R}.

} 
\examples{
\dontrun{
##vignette("Journal", package = "PMwR")}
}
\keyword{package}
