\name{REXP}
\alias{REXP}
\docType{data}
\title{
  REXP
}
\description{
    Historical Prices of the REXP.
}
\usage{data("REXP")}
\format{
  A data frame with 502 observations on the following variable:
  \describe{
    \item{\code{REXP}}{a numeric vector}
  }
}
\details{
   Daily prices.
}
%% \source{
%% %%  ~~ reference to a publication or URL from which the data were obtained ~~
%% }
%% \references{
%% %%  ~~ possibly secondary sources and usages ~~
%% }
\examples{
str(REXP)
}
\keyword{datasets}
