\name{toHTML}
\alias{toHTML}
\docType{import}
\title{Import from package \pkg{textutils}}
\description{
The \code{toHTML} function is imported from package \pkg{textutils}.
Help is available at \code{\link[textutils:toHTML]{textutils::toHTML}}.
Say \code{library("textutils")} in your code to use the function.
}
