\name{monthlyStats}
\alias{monthlyStats}
\title{
  Monthly Statistics
}
\description{
  Compute monthly returns.
}
\usage{
monthlyStats(x, t = NULL)
}
\arguments{
  \item{x}{
    a vector (or a \code{zoo} series)
  }
  \item{t}{
    time vector
  }

}
\details{
  computes an index and returns
}
\value{
  a list
  \item{index}{scaled performance}
  \item{returns}{monthly returns}

}
\author{
  Enrico Schumann
}
\examples{
##
}
