\name{DAX}
\alias{DAX}
\docType{data}
\title{
  Deutscher Aktienindex (DAX)
}
\description{
  Historical Prices of the DAX.
}
\usage{data("DAX")}
\format{
  A data frame with 505 observations on the following variable:
  \describe{
    \item{\code{DAX}}{a numeric vector}
  }
}
\details{
  Close prices.
}
%% \source{
%% %%  ~~ reference to a publication or URL from which the data were obtained ~~
%% }
%% \references{
%% %%  ~~ possibly secondary sources and usages ~~
%% }
\examples{
str(DAX)
}
\keyword{datasets}
