\name{PMwR-package}
\alias{PMwR-package}
\alias{PMwR}
\docType{package}
\title{
  Tools for the Management of Financial Portfolios
}
\description{

  Functions for the practical management of financial portfolios:
  computing profit-and-loss, accounting, analysing trades, reporting,
  and more.

}
\details{

  The package comprises functions that have been written in various
  circumstances and for various purposes.  Which is an apologetic way of
  saying that in the current version, the interfaces to functions are
  not stable (e.g., argument names are currently being made consistent
  across functions); in some cases, generic functions will be
  introduced.  The currently \bold{recommended practice} is therefore to
  \bold{explicitly name arguments in function calls} (and not pass
  arguments by position).  Any changes in argument names will be
  documented in the ChangeLog and so can be easily followed.

  The Manual provides more details. It is available from
  \url{http://enricoschumann.net/R/packages/PMwR/manual/PMwR.html} .

}
\author{
Enrico Schumann

Maintainer: Enrico Schumann <es@enricoschumann.net>
}
\references{

Gilli, M., Maringer, D. and Schumann, E. (2011) \emph{Numerical Methods
and Optimization in Finance}. Elsevier.
\url{http://www.elsevierdirect.com/product.jsp?isbn=9780123756626}

Schumann, E. (2015) \emph{Portfolio Management with R}.
\url{http://enricoschumann.net/#Books}

} 
\keyword{package}
