\name{PMwR-package}
\alias{PMwR-package}
\alias{PMwR}
\docType{package}
\title{
  Tools for the Management of Financial Portfolios
}
\description{

  Functions for the practical management of financial
  portfolios: backtesting investment and trading
  strategies, computing profit-and-loss and returns,
  analysing trades, reporting, and more.

}
\details{

\pkg{PMwR} provides a small set of reliable, efficient
and convenient tools that help in processing and
analysing trade/portfolio data. The Manual provides all
the details; it is available from
\url{http://enricoschumann.net/PMwR/}. Examples and
descriptions of new features are provided at
\url{http://enricoschumann.net/notes/PMwR/}.

}
\author{

  Enrico Schumann <es@enricoschumann.net>

}
\references{

  Schumann, E. (2022) \emph{Portfolio Management with R}.
  \url{http://enricoschumann.net/PMwR/}

  Gilli, M., Maringer, D. and Schumann, E. (2019)
  \emph{Numerical Methods and Optimization in Finance}.
  2nd edition. Elsevier.
  \url{https://www.elsevier.com/books/numerical-methods-and-optimization-in-finance/gilli/978-0-12-815065-8}

  Schumann, E. (2019) Financial Optimisation with R (\pkg{NMOF} Manual).
  \url{http://enricoschumann.net/NMOF.htm#NMOFmanual}

}
\keyword{package}
