\name{PMwR-package}
\alias{PMwR-package}
\alias{PMwR}
\docType{package}
\title{
  Tools for the Management of Financial Portfolios
}
\description{

  Tools for the practical management of financial
  portfolios: backtesting investment and trading
  strategies, computing profit-and-loss and returns,
  analysing trades, reporting, and more.

}
\details{

  \pkg{PMwR} provides a small set of reliable, efficient
  and convenient tools for processing and analysing
  trade/portfolio data. The Manual provides all the
  details; it is available from
  \url{https://enricoschumann.net/PMwR/}. Examples and
  descriptions of new features are provided at
  \url{https://enricoschumann.net/notes/PMwR/}.

}
\author{

  Enrico Schumann <es@enricoschumann.net>

}
\references{

  The \pkg{PMwR} Manual, which explains all functionality:\cr
  Schumann, E. (2023) \emph{Portfolio Management with R}.
  \url{https://enricoschumann.net/PMwR/}

  The closely-related \pkg{NMOF} package is described in:\cr
  Gilli, M., Maringer, D. and Schumann, E. (2019) \emph{Numerical Methods
    and Optimization in Finance}. 2nd edition. Elsevier.
  \doi{10.1016/C2017-0-01621-X}

  Schumann, E. (2023) Financial Optimisation with R (\pkg{NMOF} Manual).
  \url{https://enricoschumann.net/NMOF.htm#NMOFmanual}

}
\keyword{package}
