\name{PMwR-package}
\alias{PMwR-package}
\alias{PMwR}
\docType{package}
\title{
  Tools for the Management of Financial Portfolios
}
\description{

  Functions for the practical management of financial
  portfolios: backtesting investment and trading
  strategies, computing profit-and-loss and returns,
  analysing trades, reporting, and more.

}
\details{

The aim of \pkg{PMwR} is to provide a small set of
reliable, efficient and convenient tools that help in
processing and analysing trade/portfolio data. The
Manual provides all the details; it is available from
\url{http://enricoschumann.net/PMwR/}.

}
\author{

  Enrico Schumann <es@enricoschumann.net>

}
\references{

Gilli, M., Maringer, D. and Schumann, E. (2011) \emph{Numerical Methods
and Optimization in Finance}. Elsevier.
\url{https://www.elsevier.com/books/numerical-methods-and-optimization-in-finance/gilli/978-0-12-375662-6}

Schumann, E. (2019) \emph{Portfolio Management
  with R}. \url{http://enricoschumann.net/PMwR/}

}
\keyword{package}
