\name{toText}
\alias{toText}
\title{
  Convert Objects to Plain Text
}
\description{
  Converts an object into (plain) text.
}
\usage{
toText(x, \dots)

\method{toText}{default}(x, \dots)
}
\arguments{
  \item{x}{
    an object
}
  \item{\dots}{
    further arguments
}
}
\details{
x
}
\value{
  a character vector (lines of text)
}
\references{
  ES. PMwR
}
\author{
  Enrico Schumann
}
\seealso{
  \code{\link{toLatex}}, \code{\link{toHTML}}
}
\examples{
1+1
}
